pyproject.toml
[tool.poetry]
name = "pybookkeeper"
version = "0.1.0"
description = "Python app to manage your personal finances"
authors = ["Anatoly Chistov <atchistov@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "bookkeeper"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pytest-cov = "^4.0.0"
pyside6 = "^6.0.0"
python-dateutil = "^2.5.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
mypy = "^0.991"
pylint = "^2.15.10"
flake8 = "^6.0.0"
mccabe = "^0.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
